break out new component
